cleanup
updated some exercises to have consistent formatting and better task descriptions

+++ 13_MemoryManagement/Solutions/StackSize.c
@@ -1,8 +1,3 @@
-#include <stdio.h>
-#include <stdlib.h>
-#include <sys/resource.h>
-
-
 /*
 	Task:
 		a) Approximate the stack size experimentally by causing a stack overflow.
@@ -11,9 +6,14 @@
 
 
 
+#include <stdio.h>
+#include <stdlib.h>
+#include <sys/resource.h>
+
+
 int main(void)
 {
-	for (size_t i = 0; i < 1.0e8; i += .25e5)
+	for (size_t i = 0; i < 1.0e8; i += (size_t).25e5)
 	{
 		int array[i];
 		printf("no stack overflow for kB: %lu\n", i * 4 / 1000);
